refactor: extract suggestion engine generators into modular files
Extracted generateCatchSuggestions, generateGiftAndTradeSuggestions, generateBreedingSuggestions, and generateEvolutionSuggestions from suggestionEngine.ts into a new generators/ subdirectory to drastically reduce the orchestrator file size and improve AI readability.
- decoupled AssistantApiData interface to prevent circular dependencies
- migrated relevant generator constants and helpers
