+++ src/engine/assistant/suggestionEngine.ts
@@ -1,80 +1,20 @@
 import { dexDataLoader } from '../../db/DexDataLoader';
 import { pokeDB } from '../../db/PokeDB';
-import {
-  ENCOUNTER_METHOD,
-  EVO_TRIGGER,
-  type LocationAreaEncounters,
-  POKE_VERSION_MAP,
-  type PokemonMetadata,
-  type UnifiedLocation,
-} from '../../db/schema';
+import { type LocationAreaEncounters, POKE_VERSION_MAP, type PokemonMetadata } from '../../db/schema';
 import { getGenerationConfig } from '../../utils/generationConfig';
 
-const EVO_ITEM_NAMES: Record<number, string> = {
-  80: 'Sun Stone',
-  81: 'Moon Stone',
-  82: 'Fire Stone',
-  83: 'Thunder Stone',
-  84: 'Water Stone',
-  85: 'Leaf Stone',
-  198: "King's Rock",
-  210: 'Metal Coat',
-  212: 'Dragon Scale',
-  229: 'Up-Grade',
-};
-
-const POKEAPI_TO_GEN1_ITEM: Record<number, number> = {
-  81: 0x0a, // Moon Stone
-  82: 0x20, // Fire Stone
-  83: 0x21, // Thunder Stone
-  84: 0x22, // Water Stone
-  85: 0x2f, // Leaf Stone
-};
-
-const POKEAPI_TO_GEN2_ITEM: Record<number, number> = {
-  80: 0x11, // Sun Stone
-  81: 0x08, // Moon Stone
-  82: 0x16, // Fire Stone
-  83: 0x17, // Thunder Stone
-  84: 0x18, // Water Stone
-  85: 0x22, // Leaf Stone
-  198: 0x5a, // King's Rock
-  210: 0x8f, // Metal Coat
-  212: 0x82, // Dragon Scale
-  229: 0xac, // Up-Grade
-};
-
-/**
- * Maps a modern PokeAPI evolution item ID to its corresponding internal item ID
- * for a specific game generation. This is necessary because Gen 1 and Gen 2 use
- * distinct hex values for items (e.g., Moon Stone is 0x0A in Gen 1, but 0x08 in Gen 2).
- *
- * @param pokeApiId - The item ID returned from the modern PokeAPI data source.
- * @param generation - The target game generation (1 or 2).
- * @returns The internal game item ID for the given generation, or the original ID as a fallback.
- */
-function getGameItemId(pokeApiId: number, generation: number): number {
-  if (generation === 1) return POKEAPI_TO_GEN1_ITEM[pokeApiId] || pokeApiId;
-  if (generation === 2) return POKEAPI_TO_GEN2_ITEM[pokeApiId] || pokeApiId;
-  return pokeApiId;
-}
-
-import { STATIC_GIFT_DATA, STATIC_NPC_TRADE_DATA } from '../data/gen1/assistantData';
-import { getUnobtainableReason } from '../exclusives/gen1Exclusives';
-import { getGen2UnobtainableReason } from '../exclusives/gen2Exclusives';
+import { STATIC_GIFT_DATA } from '../data/gen1/assistantData';
 import type { PokemonInstance, SaveData } from '../saveParser/index';
+import type { AssistantApiData } from './generators/types';
 import type { AssistantStrategy, EncounterDetail, RejectedSuggestion, Suggestion } from './strategies/types';
 
-export interface AssistantApiData {
-  localAid: number | null;
-  localEncounters: LocationAreaEncounters[] | null;
-  missingEncounters: Record<number, LocationAreaEncounters | null>;
-  pokemonMetadata: Record<number, PokemonMetadata | null>;
-  ancestralEncounters: Record<number, Record<number, LocationAreaEncounters | null>>;
-  areaNames: Record<number, string>;
-  allLocations: UnifiedLocation[];
-}
+export type { AssistantApiData };
 
+import { generateBreedingSuggestions } from './generators/breed';
+
+import { generateCatchSuggestions } from './generators/catch';
+import { generateEvolutionSuggestions } from './generators/evolve';
+import { generateGiftAndTradeSuggestions } from './generators/trade';
 import { getStrategy } from './strategies';
 
 /**
@@ -161,35 +101,6 @@ export async function fetchAssistantApiData(saveData: SaveData, queryTargets: nu
 }
 
 /**
- * Checks if a specific bit flag is set within a continuous byte array.
- * Used primarily for validating whether in-game event flags (like catching a static encounter
- * or claiming a gift Pokémon) have been triggered in the player's save file.
- *
- * @param flags - The raw byte array extracted from the save file representing a block of event flags.
- * @param flagId - The specific zero-indexed bit ID to check.
- * @returns True if the bit is set (1), false if it is unset (0) or if the inputs are invalid.
- */
-function checkFlag(flags: Uint8Array | undefined, flagId: number | undefined) {
-  if (!flags || flagId === undefined) return false;
-  const byteIndex = flagId >> 3;
-  const bitIndex = flagId & 7;
-  const byte = flags[byteIndex];
-  if (byte === undefined) return false;
-  return (byte & (1 << bitIndex)) !== 0;
-}
-
-const METHOD_NAMES: Record<number, string> = {
-  [ENCOUNTER_METHOD.WALK]: 'walk',
-  [ENCOUNTER_METHOD.SURF]: 'surf',
-  [ENCOUNTER_METHOD.OLD_ROD]: 'old-rod',
-  [ENCOUNTER_METHOD.GOOD_ROD]: 'good-rod',
-  [ENCOUNTER_METHOD.SUPER_ROD]: 'super-rod',
-  [ENCOUNTER_METHOD.GIFT]: 'gift',
-  [ENCOUNTER_METHOD.ROCK_SMASH]: 'rock-smash',
-  [ENCOUNTER_METHOD.HEADBUTT]: 'headbutt',
-};
-
-/**
  * Core recommendation algorithm for the Assistant.
  * Aggregates data from the player's save file, Pokédex completion status, and geographic location to generate context-aware suggestions for catching, evolving, or trading Pokémon.
  *
@@ -220,145 +131,6 @@ const METHOD_NAMES: Record<number, string> = {
  */
 
 /**
- * Evaluates wild Pokémon encounters and generates actionable catch suggestions.
- *
- * This function processes two primary categories of catches:
- * 1. **Local Catch:** Pokémon that can be encountered on the exact map the player is currently on.
- *    These receive the highest base priority (120).
- * 2. **Nearby Catch:** Pokémon found within 1-8 map transitions from the player's location.
- *    Priority scales inversely with distance (closer = higher priority), leveraging the
- *    strategy's graph traversal logic.
- *
- * It mutates the provided `suggestions` array and `localPids` set to avoid redundant
- * array allocations in the hot path.
- *
- * @param apiData - Pre-fetched lookup data, containing location definitions and encounter rates.
- * @param displayVersionId - The numeric ID of the current game version (e.g. Red, Blue, Gold).
- * @param myOtIds - A set of species IDs the player physically caught (matching their Original Trainer ID), used to skip static gifts.
- * @param missingIds - A set of Pokémon IDs the player needs to obtain.
- * @param queryTargets - The top priority missing Pokémon IDs to evaluate.
- * @param saveData - The player's parsed save file, containing current location and badges.
- * @param strategy - Generation-specific logic for calculating map distances.
- * @param suggestions - The shared array where new catch suggestions are pushed.
- * @param localPids - A shared set tracking Pokémon found locally, preventing redundant nearby checks.
- */
-function generateCatchSuggestions(
-  apiData: AssistantApiData,
-  displayVersionId: number,
-  myOtIds: Set<number>,
-  missingIds: Set<number>,
-  queryTargets: number[],
-  saveData: SaveData,
-  strategy: AssistantStrategy,
-  suggestions: Suggestion[],
-  localPids: Set<number>,
-) {
-  // A. Catch logic (Local Map)
-  // Highest priority (120) is given to Pokemon found on the exact same map the player is currently standing on.
-  if (apiData.localEncounters && apiData.localEncounters.length > 0 && apiData.localAid) {
-    const localAid = apiData.localAid;
-
-    const localEncounterInfo: Record<number, EncounterDetail[]> = {};
-
-    for (const lae of apiData.localEncounters) {
-      const pid = lae.pid;
-      const relevantEncounters = lae.enc.filter((e) => e.aid === localAid && e.v === displayVersionId);
-      if (relevantEncounters.length === 0) continue;
-
-      if (STATIC_GIFT_DATA[pid] && myOtIds.has(pid)) continue;
-
-      if (missingIds.has(pid)) {
-        localPids.add(pid);
-        const details: EncounterDetail[] = [];
-        for (let r = 0; r < relevantEncounters.length; r++) {
-          const re = relevantEncounters[r];
-          if (!re) continue;
-          for (let d = 0; d < re.d.length; d++) {
-            const ed = re.d[d];
-            if (!ed) continue;
-            details.push({
-              chance: ed.c,
-              method: METHOD_NAMES[ed.m] || 'walk',
-              minLevel: ed.min,
-              maxLevel: ed.max,
-              aid: re.aid,
-              time: ed.t,
-            });
-          }
-        }
-        localEncounterInfo[pid] = details;
-      }
-    }
-
-    if (localPids.size > 0) {
-      suggestions.push({
-        id: 'catch-local',
-        category: 'Catch',
-        title: 'Catch Right Here',
-        description: `You are at ${saveData.currentMapName || 'your current location'}! There are ${localPids.size} missing Pokémon right here.`,
-        pokemonIds: Array.from(localPids),
-        priority: 120,
-        encounterInfo: localEncounterInfo,
-      });
-    }
-  }
-
-  // A2. Nearby logic (1-8 areas away)
-  // Distance is calculated via graph traversal in the generation's strategy.
-  // Priority dynamically scales inversely with distance (closer = higher priority).
-  for (const pid of queryTargets) {
-    if (localPids.has(pid)) continue;
-
-    const encData = apiData.missingEncounters[pid];
-    if (!encData?.enc) continue;
-
-    let bestDist = 999;
-    let bestAreaName = '';
-    // ⚡ Bolt: Store the best encounter reference and defer mapping EncounterDetails until after the loop
-    // to prevent redundant array allocations and O(N) mapping operations for every missing Pokémon.
-    let bestE: (typeof encData.enc)[0] | null = null;
-
-    for (const e of encData.enc) {
-      if (e.v !== displayVersionId) continue;
-
-      const distInfo = strategy.getMapDistance(saveData.currentMapId, e.aid, apiData.allLocations);
-      if (distInfo && distInfo.distance < bestDist) {
-        bestDist = distInfo.distance;
-        bestAreaName = distInfo.name;
-        bestE = e;
-      }
-    }
-
-    if (bestDist < 8 && bestE) {
-      const aid = bestE.aid;
-      const bestDetails: EncounterDetail[] = [];
-      for (let d = 0; d < bestE.d.length; d++) {
-        const ed = bestE.d[d];
-        if (!ed) continue;
-        bestDetails.push({
-          chance: ed.c,
-          method: METHOD_NAMES[ed.m] || 'walk',
-          minLevel: ed.min,
-          maxLevel: ed.max,
-          aid,
-          time: ed.t,
-        });
-      }
-
-      suggestions.push({
-        id: `catch-nearby-${pid}`,
-        category: 'Catch',
-        title: `Nearby: #${pid}`,
-        description: `Found at ${bestAreaName} (${bestDist === 0 ? 'very close' : `${bestDist} areas away`}).`,
-        pokemonId: pid,
-        priority: Math.max(10, 110 - bestDist * 12),
-        encounterInfo: { [pid]: bestDetails },
-      });
-    }
-  }
-}
-
-/**
  * Evaluates non-wild encounter methods, specifically in-game NPC trades, static gifts,
  * and version-exclusive constraints.
  *
@@ -383,133 +155,6 @@ function generateCatchSuggestions(
  */
 
 /**
- * Evaluates version exclusives, in-game NPC trades, and static gift encounters.
- *
- * It mutates the provided `suggestions` array.
- *
- * @param queryTargets - The top priority missing Pokémon IDs to evaluate.
- * @param saveData - The player's parsed save file, containing badges and event flags.
- * @param displayVersion - The current game version string.
- * @param ownedSet - A Set of Pokémon IDs the player already owns.
- * @param apiData - Pre-fetched metadata for Pokémon definitions.
- * @param instancesBySpecies - A Map of the player's physical Pokémon, used to check for required trade offerings or pre-evolutions.
- * @param suggestions - The shared array where new suggestions are pushed.
- * @param missingIds - A Set of Pokémon IDs the player needs to obtain.
- */
-function generateGiftAndTradeSuggestions(
-  queryTargets: number[],
-  saveData: SaveData,
-  displayVersion: string,
-  ownedSet: Set<number>,
-  apiData: AssistantApiData,
-  instancesBySpecies: Map<number, PokemonInstance[]>,
-  suggestions: Suggestion[],
-  missingIds: Set<number>,
-) {
-  // B. Unobtainable / Exclusive logic
-  // Checks if the target is completely locked out of the current version (e.g. Red exclusives on Blue).
-  // These are assigned the lowest base priority (10) since they require external action (link cable trades).
-  // ⚡ Bolt: Convert O(N^2) array lookup to O(1) Set has for NPC trades
-  const validNpcTradeIds = new Set<number>();
-  for (let i = 0; i < STATIC_NPC_TRADE_DATA.length; i++) {
-    const t = STATIC_NPC_TRADE_DATA[i];
-    if (t && t.gen === saveData.generation && (!t.versions || t.versions.includes(displayVersion))) {
-      validNpcTradeIds.add(t.receivedId);
-    }
-  }
-
-  const pidsWithExclusives = new Set<number>();
-  for (const pid of queryTargets) {
-    let reason: string | null = null;
-    if (saveData.generation === 2) {
-      reason = getGen2UnobtainableReason(pid, displayVersion, ownedSet.size, ownedSet);
-    } else {
-      reason = getUnobtainableReason(pid, displayVersion, ownedSet.size, ownedSet);
-    }
-    if (reason) {
-      pidsWithExclusives.add(pid);
-
-      const isNpcTrade = validNpcTradeIds.has(pid);
-      if (isNpcTrade) continue;
-
-      // If they physically own a pre-evolution, they don't strictly need to trade, they can evolve it!
-      const p = apiData.pokemonMetadata?.[pid];
-      let hasPhysicalPreEvo = false;
-      if (p?.efrm && p.efrm.length > 0) {
-        // Iterate backwards through all ancestors (recursive ownership check)
-        // The current logic only checked immediate parents, so if a player had Charmander,
-        // Charizard might incorrectly be flagged as Unobtainable/Trade.
-        for (let i = p.efrm.length - 1; i >= 0; i--) {
-          const preId = p.efrm[i];
-          if (preId !== undefined && instancesBySpecies.has(preId)) {
-            hasPhysicalPreEvo = true;
-            break;
-          }
-        }
-      }
-      if (hasPhysicalPreEvo) continue;
-
-      suggestions.push({
-        id: `exclusive-${pid}`,
-        category: 'Trade',
-        title: `Version Exclusive: #${pid}`,
-        description: reason,
-        pokemonId: pid,
-        priority: 10,
-      });
-    }
-  }
-
-  // C. In-Game NPC Trades
-  // Priority boosts if the player already physically possesses the required "offered" Pokemon (65 -> 85).
-  for (const trade of STATIC_NPC_TRADE_DATA) {
-    if (trade.gen !== saveData.generation) continue;
-    if (trade.versions && !trade.versions.includes(displayVersion)) continue;
-    if (!missingIds.has(trade.receivedId)) continue;
-
-    if (trade.tradeIndex !== undefined && saveData.npcTradeFlags !== undefined) {
-      const isClaimed = (saveData.npcTradeFlags & (1 << trade.tradeIndex)) !== 0;
-      if (isClaimed) continue;
-    }
-
-    const hasOffered = instancesBySpecies.has(trade.offeredId);
-    suggestions.push({
-      id: `npc-trade-${trade.receivedId}`,
-      category: 'Trade',
-      title: `Trade for #${trade.receivedId}`,
-      description: hasOffered
-        ? `You have #${trade.offeredId}! Trade it at ${trade.location} for #${trade.receivedId}.`
-        : `Catch #${trade.offeredId} and trade it at ${trade.location} for #${trade.receivedId}.`,
-      pokemonId: trade.receivedId,
-      priority: hasOffered ? 85 : 65,
-    });
-  }
-
-  // D. Static Gifts
-  // Suggests available static encounters and gifts that haven't been claimed yet.
-  for (const [idStr, gift] of Object.entries(STATIC_GIFT_DATA)) {
-    const giftId = parseInt(idStr, 10);
-    if (gift.gen && gift.gen !== saveData.generation) continue;
-    if (!missingIds.has(giftId)) continue;
-
-    const requiredBadges = gift.requiredBadges || 0;
-    if (saveData.badges < requiredBadges) continue;
-
-    const hasClaimed = checkFlag(saveData.eventFlags, gift.eventFlag);
-    if (hasClaimed) continue;
-
-    suggestions.push({
-      id: `gift-${giftId}`,
-      category: 'Gift',
-      title: `Claim Gift: #${giftId}`,
-      description: `Get ${gift.name} at ${gift.location} (${gift.reason}).`,
-      pokemonId: giftId,
-      priority: 85,
-    });
-  }
-}
-
-/**
  * Evaluates Gen 2 Daycare breeding logic.
  * Checks if the player can breed a missing base Pokémon from an owned evolution.
  *
@@ -521,82 +166,6 @@ function generateGiftAndTradeSuggestions(
  * @param instancesBySpecies - A Map of the player's physical Pokémon.
  * @param suggestions - The shared array where new breeding suggestions are pushed.
  */
-function generateBreedingSuggestions(
-  queryTargets: number[],
-  saveData: SaveData,
-  apiData: AssistantApiData,
-  instancesBySpecies: Map<number, PokemonInstance[]>,
-  suggestions: Suggestion[],
-) {
-  // F. Breeding (Gen 2 Only)
-  if (saveData.generation === 2) {
-    queryTargets.forEach((targetId: number) => {
-      const p = apiData.pokemonMetadata?.[targetId];
-      if (!p) return;
-
-      // Check if we are missing a base Pokemon, but we own an evolution of it
-      let canBreed = false;
-      let evolutionIdToBreed: number | null = null;
-
-      // Only base Pokemon can be hatched from an egg
-      if (p.efrm === undefined || p.efrm.length === 0) {
-        // Look at all evolutions of the target (recursive)
-        const stack = [...(p.eto || [])];
-        while (stack.length > 0) {
-          const evo = stack.pop();
-          if (
-            evo &&
-            (instancesBySpecies.has(evo.id) || (saveData.daycare?.some((d) => d.speciesId === evo.id) ?? false))
-          ) {
-            canBreed = true;
-            evolutionIdToBreed = evo.id;
-            break;
-          }
-          if (evo?.eto && evo.eto.length > 0) {
-            stack.push(...evo.eto);
-          }
-        }
-      }
-
-      if (canBreed && evolutionIdToBreed) {
-        const isInDaycare = saveData.daycare?.some((d) => d.speciesId === evolutionIdToBreed) ?? false;
-
-        let description = `Leave your #${evolutionIdToBreed} at the Daycare to get an Egg!`;
-        let priority = 85;
-        let title = `Breed: #${targetId}`;
-
-        if (isInDaycare) {
-          if (saveData.daycare && saveData.daycare.length === 2) {
-            if (saveData.daycareHasEgg) {
-              title = `Egg Ready: #${targetId}!`;
-              description = `Pick up your Egg from the Daycare!`;
-              priority = 95;
-            } else {
-              title = `Breeding in Progress: #${targetId}`;
-              description = `Wait for an Egg from the Daycare!`;
-              priority = 85;
-            }
-          } else {
-            title = `Need Partner: #${targetId}`;
-            description = `Leave a compatible partner (like Ditto) at the Daycare to get an Egg!`;
-            priority = 80;
-          }
-        } else {
-          description = `Leave your #${evolutionIdToBreed} and a compatible partner (like Ditto) at the Daycare to get an Egg!`;
-        }
-
-        suggestions.push({
-          id: `breed-${targetId}`,
-          category: 'Breed',
-          title,
-          description,
-          pokemonId: targetId,
-          priority,
-        });
-      }
-    });
-  }
-}
 
 /**
  * Evaluates the player's current boxes and party to find pre-evolutions that can be evolved
@@ -616,226 +185,6 @@ function generateBreedingSuggestions(
  * @param displayVersion - The current game version, used to handle special cases (like Yellow Pikachu refusing to evolve).
  * @param missingIds - A Set of Pokémon IDs the player needs to obtain.
  */
-function generateEvolutionSuggestions(
-  queryTargets: number[],
-  saveData: SaveData,
-  apiData: AssistantApiData,
-  instancesBySpecies: Map<number, PokemonInstance[]>,
-  suggestions: Suggestion[],
-  displayVersion: string,
-  missingIds: Set<number>,
-) {
-  // E. Evolutions
-  // Evaluates the player's current boxes and party to find pre-evolutions.
-  // Priority boosts significantly if the evolution criteria are actively met (e.g. required level reached, evolution stone in inventory).
-
-  queryTargets.forEach((targetId: number) => {
-    const p = apiData.pokemonMetadata?.[targetId];
-    if (!p) return;
-
-    let closestOwnedParentId: number | undefined;
-    let immediateEvoTargetId: number = targetId;
-
-    for (let i = 0; i < p.efrm.length; i++) {
-      const ancestorId = p.efrm[i];
-      if (ancestorId !== undefined && instancesBySpecies.has(ancestorId)) {
-        closestOwnedParentId = ancestorId;
-        const nextTarget = i === 0 ? targetId : p.efrm[i - 1];
-        if (nextTarget !== undefined) {
-          immediateEvoTargetId = nextTarget;
-        }
-        break;
-      }
-    }
-
-    if (closestOwnedParentId === undefined) return;
-    const ownedInstances = instancesBySpecies.get(closestOwnedParentId) || [];
-    if (ownedInstances.length === 0) return;
-
-    const immediateEvoTarget = apiData.pokemonMetadata?.[immediateEvoTargetId];
-    if (!immediateEvoTarget) return;
-
-    // If we're looking at a multi-stage evolution (e.g., target is Charizard, immediate is Charmeleon)
-    // AND the intermediate stage (Charmeleon) is ALSO missing from the Pokedex,
-    // we should skip generating the suggestion for the final stage (Charizard) because
-    // the engine will already generate an identical "Evolve Charmander -> Charmeleon" suggestion
-    // when it evaluates Charmeleon as a target. This prevents redundant duplicates.
-    if (immediateEvoTargetId !== targetId && missingIds.has(immediateEvoTargetId)) {
-      return;
-    }
-
-    const details = immediateEvoTarget.det;
-    if (!details || details.length === 0) return;
-
-    for (const detail of details) {
-      const tr = detail.tr;
-      const min_l = detail.ml;
-      const min_h = detail.mh;
-      const item = detail.item;
-      const held = detail.held;
-      const tod = detail.time === 1 ? 'day' : detail.time === 2 ? 'night' : undefined;
-      const rps = detail.rps;
-
-      // Filter out Yellow Starter Pikachu as it refuses to evolve
-      const evolvableInstances = ownedInstances.filter(
-        (inst) => !(displayVersion === 'yellow' && closestOwnedParentId === 25 && inst.otName === saveData.trainerName),
-      );
-
-      if (evolvableInstances.length === 0) continue;
-
-      let bestInstance = evolvableInstances[0];
-      if (!bestInstance) continue;
-      if (tr === EVO_TRIGGER.LEVEL_UP && min_h) {
-        bestInstance = evolvableInstances.reduce((prev, current) =>
-          (prev.friendship ?? 0) > (current.friendship ?? 0) ? prev : current,
-        );
-      } else {
-        bestInstance = evolvableInstances.reduce((prev, current) => (prev.level > current.level ? prev : current));
-      }
-
-      const isIntermediate = immediateEvoTargetId !== targetId;
-      const pathTitlePrefix = isIntermediate ? `Path to #${targetId}` : 'Evolution';
-      const evolveTargetText = isIntermediate ? ` into #${immediateEvoTargetId} to progress towards #${targetId}` : '';
-
-      if (tr === EVO_TRIGGER.LEVEL_UP) {
-        if (min_l) {
-          const isReady = bestInstance.level >= min_l;
-          let rpsMet = true;
-          if (rps !== undefined && bestInstance.dvs && bestInstance.statExp) {
-            const baseAtk = 35;
-            const baseDef = 35;
-            const calcAtk =
-              Math.floor(
-                (((baseAtk + bestInstance.dvs.atk) * 2 +
-                  Math.floor(Math.min(Math.floor(Math.ceil(Math.sqrt(bestInstance.statExp.atk))), 255) / 4)) *
-                  bestInstance.level) /
-                  100,
-              ) + 5;
-            const calcDef =
-              Math.floor(
-                (((baseDef + bestInstance.dvs.def) * 2 +
-                  Math.floor(Math.min(Math.floor(Math.ceil(Math.sqrt(bestInstance.statExp.def))), 255) / 4)) *
-                  bestInstance.level) /
-                  100,
-              ) + 5;
-            if (rps === 1) rpsMet = calcAtk > calcDef;
-            else if (rps === -1) rpsMet = calcAtk < calcDef;
-            else if (rps === 0) rpsMet = calcAtk === calcDef;
-          }
-          const isActuallyReady = isReady && rpsMet;
-          let rpsReq = '';
-          if (rps === 1) rpsReq = ', Atk > Def';
-          else if (rps === -1) rpsReq = ', Atk < Def';
-          else if (rps === 0) rpsReq = ', Atk = Def';
-          let specificReq = `(needs Lv. ${min_l}${rpsReq})`;
-          if (!rpsReq) specificReq = `(needs Lv. ${min_l})`;
-
-          suggestions.push({
-            id: `evo-lvl-${targetId}`,
-            category: 'Evolve',
-            title: isIntermediate ? pathTitlePrefix : `Level Up Evolution: #${targetId}`,
-            description: isActuallyReady
-              ? `Your Lv. ${bestInstance.level} pre-evolution is ready to evolve${evolveTargetText} ${specificReq}!`
-              : `Your Lv. ${bestInstance.level} pre-evolution evolves at Lv. ${min_l}${evolveTargetText} ${specificReq}.`,
-            pokemonId: targetId,
-            priority: isActuallyReady ? 90 : 75,
-          });
-        } else if (min_h) {
-          const todMsg = tod ? ` during the ${tod}` : '';
-          const isFriendlyEnough = bestInstance.friendship !== undefined && bestInstance.friendship >= min_h;
-          const friendshipStatus =
-            bestInstance.friendship !== undefined ? ` (${bestInstance.friendship}/${min_h})` : '';
-
-          suggestions.push({
-            id: `evo-happy-${targetId}`,
-            category: 'Evolve',
-            title: isIntermediate
-              ? pathTitlePrefix
-              : isFriendlyEnough
-                ? `Ready to Evolve: #${targetId}!`
-                : `Happiness Evolution: #${targetId}`,
-            description: isFriendlyEnough
-              ? `Your pre-evolution is friendly enough${friendshipStatus}! Level it up${todMsg} to evolve${evolveTargetText}.`
-              : `Level up your pre-evolution with high happiness${friendshipStatus} to evolve${todMsg}${evolveTargetText}!`,
-            pokemonId: targetId,
-            priority: isFriendlyEnough ? 90 : 80,
-          });
-        } else {
-          const todMsg = tod ? ` during the ${tod}` : '';
-          suggestions.push({
-            id: `evo-lvl-any-${targetId}`,
-            category: 'Evolve',
-            title: isIntermediate ? pathTitlePrefix : `Level Up Evolution: #${targetId}`,
-            description: `Level up your pre-evolution${todMsg} to evolve${evolveTargetText}!`,
-            pokemonId: targetId,
-            priority: 70,
-          });
-        }
-      } else if (tr === EVO_TRIGGER.USE_ITEM && item) {
-        const gameItemId = getGameItemId(item, saveData.generation);
-        const hasStone =
-          saveData.inventory.some((i) => i.id === gameItemId && i.quantity > 0) ||
-          (saveData.pcItems?.some((i) => i.id === gameItemId && i.quantity > 0) ?? false);
-        const itemName = EVO_ITEM_NAMES[item] || 'item';
-        suggestions.push({
-          id: `evo-item-${targetId}-${item}`,
-          category: 'Evolve',
-          title: isIntermediate
-            ? pathTitlePrefix
-            : hasStone
-              ? `Ready to Evolve: #${targetId}!`
-              : `Item Needed: #${targetId}`,
-          description: hasStone
-            ? `Use your ${itemName} to evolve it${evolveTargetText}!`
-            : `Find a ${itemName} to evolve it${evolveTargetText}.`,
-          pokemonId: targetId,
-          priority: hasStone ? 95 : 40,
-        });
-      } else if (tr === EVO_TRIGGER.TRADE) {
-        if (held) {
-          const gameHeldId = getGameItemId(held, saveData.generation);
-          const hasHeldItemInBag =
-            saveData.inventory.some((i) => i.id === gameHeldId && i.quantity > 0) ||
-            (saveData.pcItems?.some((i) => i.id === gameHeldId && i.quantity > 0) ?? false);
-          const holdingInstance =
-            evolvableInstances.find((inst) => inst.item === gameHeldId) ||
-            ownedInstances.find((inst) => inst.item === gameHeldId);
-          const hasHeldItem = hasHeldItemInBag || !!holdingInstance;
-          const itemName = EVO_ITEM_NAMES[held] || 'item';
-
-          let description = `Find a ${itemName}, have your pre-evolution hold it, and trade to evolve${evolveTargetText}.`;
-          if (holdingInstance) {
-            description = `Your pre-evolution is already holding the ${itemName}! Trade it to evolve${evolveTargetText}!`;
-          } else if (hasHeldItemInBag) {
-            description = `Have your pre-evolution hold the ${itemName} and trade it to evolve${evolveTargetText}!`;
-          }
-
-          suggestions.push({
-            id: `evo-trade-held-${targetId}`,
-            category: 'Evolve',
-            title: isIntermediate
-              ? pathTitlePrefix
-              : hasHeldItem
-                ? `Ready to Trade Evolve: #${targetId}!`
-                : `Item Needed for Trade: #${targetId}`,
-            description,
-            pokemonId: targetId,
-            priority: hasHeldItem ? 90 : 45,
-          });
-        } else {
-          suggestions.push({
-            id: `evo-trade-${targetId}`,
-            category: 'Evolve',
-            title: isIntermediate ? pathTitlePrefix : `Trade Evolution: #${targetId}`,
-            description: `Trade your pre-evolution to evolve it${evolveTargetText}!`,
-            pokemonId: targetId,
-            priority: 85,
-          });
-        }
-      }
-    }
-  });
-}
 
 /**
  * The core orchestration function for the Assistant recommendation engine.
